Reorganized Async & Sync OperationHelpers
Normalized some naming conventions (mainly async oddities)
Removed any unused code

JAVA-5086
