+++ driver-core/src/main/com/mongodb/internal/operation/AbortTransactionOperation.java
@@ -21,6 +21,8 @@ import com.mongodb.WriteConcern;
 import com.mongodb.lang.Nullable;
 import org.bson.BsonDocument;
 
+import static com.mongodb.internal.operation.CommandOperationHelper.CommandCreator;
+
 /**
  * An operation that aborts a transaction.
  *
@@ -44,8 +46,8 @@ public class AbortTransactionOperation extends TransactionOperation {
     }
 
     @Override
-    CommandOperationHelper.CommandCreator getCommandCreator() {
-        CommandOperationHelper.CommandCreator creator = super.getCommandCreator();
+    CommandCreator getCommandCreator() {
+        CommandCreator creator = super.getCommandCreator();
         if (recoveryToken != null) {
             return (serverDescription, connectionDescription) -> creator.create(serverDescription, connectionDescription).append("recoveryToken", recoveryToken);
         }
