+++ driver-core/src/main/com/mongodb/internal/operation/BaseFindAndModifyOperation.java
@@ -36,13 +36,13 @@ import org.bson.conversions.Bson;
 import java.util.concurrent.TimeUnit;
 
 import static com.mongodb.assertions.Assertions.notNull;
+import static com.mongodb.internal.operation.AsyncOperationHelper.executeRetryableWriteAsync;
 import static com.mongodb.internal.operation.CommandOperationHelper.CommandCreator;
-import static com.mongodb.internal.operation.CommandOperationHelper.executeRetryableWrite;
-import static com.mongodb.internal.operation.CommandOperationHelper.executeRetryableWriteAsync;
 import static com.mongodb.internal.operation.DocumentHelper.putIfNotNull;
 import static com.mongodb.internal.operation.DocumentHelper.putIfNotZero;
 import static com.mongodb.internal.operation.OperationHelper.isRetryableWrite;
 import static com.mongodb.internal.operation.OperationHelper.validateHintForFindAndModify;
+import static com.mongodb.internal.operation.SyncOperationHelper.executeRetryableWrite;
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
 
 /**
