+++ driver-core/src/main/com/mongodb/internal/operation/CommandReadOperation.java
@@ -23,9 +23,9 @@ import org.bson.BsonDocument;
 import org.bson.codecs.Decoder;
 
 import static com.mongodb.assertions.Assertions.notNull;
+import static com.mongodb.internal.operation.AsyncOperationHelper.executeRetryableReadAsync;
 import static com.mongodb.internal.operation.CommandOperationHelper.CommandCreator;
-import static com.mongodb.internal.operation.CommandOperationHelper.executeRetryableRead;
-import static com.mongodb.internal.operation.CommandOperationHelper.executeRetryableReadAsync;
+import static com.mongodb.internal.operation.SyncOperationHelper.executeRetryableRead;
 
 /**
  * An operation that executes an arbitrary command that reads from the server.
