+++ driver-core/src/main/com/mongodb/internal/operation/OperationHelper.java
@@ -24,23 +24,12 @@ import com.mongodb.client.model.Collation;
 import com.mongodb.connection.ConnectionDescription;
 import com.mongodb.connection.ServerDescription;
 import com.mongodb.connection.ServerType;
-import com.mongodb.internal.async.AsyncBatchCursor;
 import com.mongodb.internal.async.SingleResultCallback;
-import com.mongodb.internal.async.function.AsyncCallbackBiFunction;
 import com.mongodb.internal.async.function.AsyncCallbackFunction;
 import com.mongodb.internal.async.function.AsyncCallbackSupplier;
-import com.mongodb.internal.binding.AsyncConnectionSource;
-import com.mongodb.internal.binding.AsyncReadBinding;
-import com.mongodb.internal.binding.AsyncWriteBinding;
-import com.mongodb.internal.binding.ConnectionSource;
-import com.mongodb.internal.binding.ReadBinding;
-import com.mongodb.internal.binding.ReferenceCounted;
-import com.mongodb.internal.binding.WriteBinding;
 import com.mongodb.internal.bulk.DeleteRequest;
 import com.mongodb.internal.bulk.UpdateRequest;
 import com.mongodb.internal.bulk.WriteRequest;
-import com.mongodb.internal.connection.AsyncConnection;
-import com.mongodb.internal.connection.Connection;
 import com.mongodb.internal.connection.QueryResult;
 import com.mongodb.internal.diagnostics.logging.Logger;
 import com.mongodb.internal.diagnostics.logging.Loggers;
@@ -49,46 +38,21 @@ import com.mongodb.lang.NonNull;
 import com.mongodb.lang.Nullable;
 import org.bson.BsonDocument;
 import org.bson.BsonInt64;
-import org.bson.BsonValue;
 import org.bson.codecs.Decoder;
 import org.bson.conversions.Bson;
 
 import java.util.Collections;
 import java.util.List;
-import java.util.function.BiFunction;
 import java.util.function.Function;
 import java.util.function.Supplier;
 
 import static com.mongodb.assertions.Assertions.assertNotNull;
-import static com.mongodb.assertions.Assertions.notNull;
-import static com.mongodb.internal.async.ErrorHandlingResultCallback.errorHandlingCallback;
 import static com.mongodb.internal.operation.ServerVersionHelper.serverIsLessThanVersionFourDotFour;
 import static com.mongodb.internal.operation.ServerVersionHelper.serverIsLessThanVersionFourDotTwo;
 import static java.lang.String.format;
-import static java.util.Collections.singletonList;
 
 final class OperationHelper {
     public static final Logger LOGGER = Loggers.getLogger("operation");
-
-    interface CallableWithConnection<T> {
-        T call(Connection connection);
-    }
-
-    interface CallableWithSource<T> {
-        T call(ConnectionSource source);
-    }
-
-    interface AsyncCallableWithConnection {
-        void call(@Nullable AsyncConnection connection, @Nullable Throwable t);
-    }
-
-    interface AsyncCallableWithSource {
-        void call(@Nullable AsyncConnectionSource source, @Nullable Throwable t);
-    }
-
-    interface AsyncCallableWithConnectionAndSource {
-        void call(@Nullable AsyncConnectionSource source, @Nullable AsyncConnection connection, @Nullable Throwable t);
-    }
 
     static void validateCollationAndWriteConcern(@Nullable final Collation collation, final WriteConcern writeConcern) {
         if (collation != null && !writeConcern.isAcknowledged()) {
@@ -127,7 +91,7 @@ final class OperationHelper {
         }
     }
 
-    static void validateWriteRequestCollations(final List<? extends WriteRequest> requests, final WriteConcern writeConcern) {
+    private static void validateWriteRequestCollations(final List<? extends WriteRequest> requests, final WriteConcern writeConcern) {
         Collation collation = null;
         for (WriteRequest request : requests) {
             if (request instanceof UpdateRequest) {
@@ -142,7 +106,7 @@ final class OperationHelper {
         validateCollationAndWriteConcern(collation, writeConcern);
     }
 
-    static void validateUpdateRequestArrayFilters(final List<? extends WriteRequest> requests, final WriteConcern writeConcern) {
+    private static void validateUpdateRequestArrayFilters(final List<? extends WriteRequest> requests, final WriteConcern writeConcern) {
         for (WriteRequest request : requests) {
             List<BsonDocument> arrayFilters = null;
             if (request instanceof UpdateRequest) {
@@ -155,7 +119,8 @@ final class OperationHelper {
         }
     }
 
-    static void validateWriteRequestHints(final ConnectionDescription connectionDescription, final List<? extends WriteRequest> requests,
+    private static void validateWriteRequestHints(final ConnectionDescription connectionDescription,
+            final List<? extends WriteRequest> requests,
             final WriteConcern writeConcern) {
         for (WriteRequest request : requests) {
             Bson hint = null;
@@ -194,7 +159,7 @@ final class OperationHelper {
         }
     }
 
-    static void checkBypassDocumentValidationIsSupported(@Nullable final Boolean bypassDocumentValidation,
+    private static void checkBypassDocumentValidationIsSupported(@Nullable final Boolean bypassDocumentValidation,
             final WriteConcern writeConcern) {
         if (bypassDocumentValidation != null && !writeConcern.isAcknowledged()) {
             throw new MongoClientException("Specifying bypassDocumentValidation with an unacknowledged WriteConcern is not supported");
@@ -242,277 +207,36 @@ final class OperationHelper {
                 0, batchSize, decoder);
     }
 
-    static <T> AsyncBatchCursor<T> createEmptyAsyncBatchCursor(final MongoNamespace namespace, final ServerAddress serverAddress) {
-        return new AsyncSingleBatchQueryCursor<>(new QueryResult<>(namespace, Collections.emptyList(), 0L, serverAddress));
-    }
-
-    static <T> BatchCursor<T> cursorDocumentToBatchCursor(final BsonDocument cursorDocument, final Decoder<T> decoder,
-            final BsonValue comment, final ConnectionSource source, final Connection connection, final int batchSize) {
-        return new QueryBatchCursor<>(OperationHelper.cursorDocumentToQueryResult(cursorDocument,
-                source.getServerDescription().getAddress()),
-                0, batchSize, 0, decoder, comment, source, connection);
-    }
-
-    static <T> AsyncBatchCursor<T> cursorDocumentToAsyncBatchCursor(final BsonDocument cursorDocument, final Decoder<T> decoder,
-            final BsonValue comment, final AsyncConnectionSource source, final AsyncConnection connection, final int batchSize) {
-        return new AsyncQueryBatchCursor<>(OperationHelper.cursorDocumentToQueryResult(cursorDocument,
-                source.getServerDescription().getAddress()),
-                0, batchSize, 0, decoder, comment, source, connection, cursorDocument);
-    }
-
-
     static <T> QueryResult<T> cursorDocumentToQueryResult(final BsonDocument cursorDocument, final ServerAddress serverAddress) {
         return cursorDocumentToQueryResult(cursorDocument, serverAddress, "firstBatch");
     }
 
-    static <T> QueryResult<T> getMoreCursorDocumentToQueryResult(final BsonDocument cursorDocument, final ServerAddress serverAddress) {
-        return cursorDocumentToQueryResult(cursorDocument, serverAddress, "nextBatch");
-    }
-
-    private static <T> QueryResult<T> cursorDocumentToQueryResult(final BsonDocument cursorDocument, final ServerAddress serverAddress,
-                                                                  final String fieldNameContainingBatch) {
+    static <T> QueryResult<T> cursorDocumentToQueryResult(final BsonDocument cursorDocument, final ServerAddress serverAddress,
+            final String fieldNameContainingBatch) {
         long cursorId = ((BsonInt64) cursorDocument.get("id")).getValue();
         MongoNamespace queryResultNamespace = new MongoNamespace(cursorDocument.getString("ns").getValue());
         return new QueryResult<>(queryResultNamespace, BsonDocumentWrapperHelper.toList(cursorDocument, fieldNameContainingBatch),
                 cursorId, serverAddress);
     }
 
-    static <T> SingleResultCallback<T> releasingCallback(final SingleResultCallback<T> wrapped, final AsyncConnection connection) {
-        return new ReferenceCountedReleasingWrappedCallback<>(wrapped, singletonList(connection));
-    }
-
-    private static class ReferenceCountedReleasingWrappedCallback<T> implements SingleResultCallback<T> {
-        private final SingleResultCallback<T> wrapped;
-        private final List<? extends ReferenceCounted> referenceCounted;
-
-        ReferenceCountedReleasingWrappedCallback(final SingleResultCallback<T> wrapped,
-                                                 final List<? extends ReferenceCounted> referenceCounted) {
-            this.wrapped = wrapped;
-            this.referenceCounted = notNull("referenceCounted", referenceCounted);
-        }
-
-        @Override
-        public void onResult(@Nullable final T result, @Nullable  final Throwable t) {
-            for (ReferenceCounted cur : referenceCounted) {
-                if (cur != null) {
-                    cur.release();
-                }
-            }
-            wrapped.onResult(result, t);
-        }
-    }
-
-    static <T> T withReadConnectionSource(final ReadBinding binding, final CallableWithSource<T> callable) {
-        ConnectionSource source = binding.getReadConnectionSource();
-        try {
-            return callable.call(source);
-        } finally {
-            source.release();
-        }
-    }
-
-    static <T> T withConnection(final WriteBinding binding, final CallableWithConnection<T> callable) {
-        ConnectionSource source = binding.getWriteConnectionSource();
-        try {
-            return withConnectionSource(source, callable);
-        } finally {
-            source.release();
-        }
-    }
-
-    static <T> T withConnectionSource(final ConnectionSource source, final CallableWithConnection<T> callable) {
-        Connection connection = source.getConnection();
-        try {
-            return callable.call(connection);
-        } finally {
-            connection.release();
-        }
-    }
-
-    /**
-     * Gets a {@link ConnectionSource} and a {@link Connection} from the {@code sourceSupplier} and executes the {@code function} with them.
-     * Guarantees to {@linkplain ReferenceCounted#release() release} the source and the connection after completion of the {@code function}.
-     *
-     * @param wrapSourceConnectionException See {@link #withSuppliedResource(Supplier, boolean, Function)}.
-     * @see #withSuppliedResource(Supplier, boolean, Function)
-     * @see #withAsyncSourceAndConnection(AsyncCallbackSupplier, boolean, SingleResultCallback, AsyncCallbackBiFunction)
-     */
-    static <R> R withSourceAndConnection(final Supplier<ConnectionSource> sourceSupplier,
-            final boolean wrapSourceConnectionException,
-            final BiFunction<ConnectionSource, Connection, R> function) throws ResourceSupplierInternalException {
-        return withSuppliedResource(sourceSupplier, wrapSourceConnectionException, source ->
-                withSuppliedResource(source::getConnection, wrapSourceConnectionException, connection ->
-                        function.apply(source, connection)));
-    }
-
-    /**
-     * Gets a {@link ReferenceCounted} resource from the {@code resourceSupplier} and applies the {@code function} to it.
-     * Guarantees to {@linkplain ReferenceCounted#release() release} the resource after completion of the {@code function}.
-     *
-     * @param wrapSupplierException If {@code true} and {@code resourceSupplier} completes abruptly, then the exception is wrapped
-     * into {@link ResourceSupplierInternalException}, such that it can be accessed
-     * via {@link ResourceSupplierInternalException#getCause()}.
-     * @see #withAsyncSuppliedResource(AsyncCallbackSupplier, boolean, SingleResultCallback, AsyncCallbackFunction)
-     */
-    static <R, T extends ReferenceCounted> R withSuppliedResource(final Supplier<T> resourceSupplier,
-            final boolean wrapSupplierException, final Function<T, R> function) throws ResourceSupplierInternalException {
-        T resource = null;
-        try {
-            try {
-                resource = resourceSupplier.get();
-            } catch (Exception supplierException) {
-                if (wrapSupplierException) {
-                    throw new ResourceSupplierInternalException(supplierException);
-                } else {
-                    throw supplierException;
-                }
-            }
-            return function.apply(resource);
-        } finally {
-            if (resource != null) {
-                resource.release();
-            }
-        }
-    }
-
-    static void withAsyncConnection(final AsyncWriteBinding binding, final AsyncCallableWithConnection callable) {
-        binding.getWriteConnectionSource(errorHandlingCallback(new AsyncCallableWithConnectionCallback(callable), LOGGER));
-    }
-
-    static void withAsyncConnection(final AsyncWriteBinding binding, final AsyncCallableWithConnectionAndSource callable) {
-        binding.getWriteConnectionSource(errorHandlingCallback(new AsyncCallableWithConnectionAndSourceCallback(callable), LOGGER));
-    }
-
-    static void withAsyncReadConnection(final AsyncReadBinding binding, final AsyncCallableWithSource callable) {
-        binding.getReadConnectionSource(errorHandlingCallback(new AsyncCallableWithSourceCallback(callable), LOGGER));
-    }
-
-    /**
-     * @see #withAsyncSuppliedResource(AsyncCallbackSupplier, boolean, SingleResultCallback, AsyncCallbackFunction)
-     * @see #withSourceAndConnection(Supplier, boolean, BiFunction)
-     */
-    static <R> void withAsyncSourceAndConnection(final AsyncCallbackSupplier<AsyncConnectionSource> sourceAsyncSupplier,
-            final boolean wrapSourceConnectionException,
-            final SingleResultCallback<R> callback,
-            final AsyncCallbackBiFunction<AsyncConnectionSource, AsyncConnection, R> asyncFunction)
-            throws ResourceSupplierInternalException {
-        SingleResultCallback<R> errorHandlingCallback = errorHandlingCallback(callback, LOGGER);
-        withAsyncSuppliedResource(sourceAsyncSupplier, wrapSourceConnectionException, errorHandlingCallback,
-                (source, sourceReleasingCallback) ->
-                        withAsyncSuppliedResource(source::getConnection, wrapSourceConnectionException, sourceReleasingCallback,
-                                (connection, connectionAndSourceReleasingCallback) ->
-                                        asyncFunction.apply(source, connection, connectionAndSourceReleasingCallback)));
-    }
-
-    /**
-     * @see #withSuppliedResource(Supplier, boolean, Function)
-     */
-    static <R, T extends ReferenceCounted> void withAsyncSuppliedResource(final AsyncCallbackSupplier<T> resourceSupplier,
-            final boolean wrapSourceConnectionException, final SingleResultCallback<R> callback,
-            final AsyncCallbackFunction<T, R> function) throws ResourceSupplierInternalException {
-        SingleResultCallback<R> errorHandlingCallback = errorHandlingCallback(callback, LOGGER);
-        resourceSupplier.get((resource, supplierException) -> {
-            if (supplierException != null) {
-                if (wrapSourceConnectionException) {
-                    supplierException = new ResourceSupplierInternalException(supplierException);
-                }
-                errorHandlingCallback.onResult(null, supplierException);
-            } else {
-                assertNotNull(resource);
-                AsyncCallbackSupplier<R> curriedFunction = clbk -> function.apply(resource, clbk);
-                curriedFunction.whenComplete(resource::release)
-                        .get(errorHandlingCallback);
-            }
-        });
-    }
-
-    private static class AsyncCallableWithConnectionCallback implements SingleResultCallback<AsyncConnectionSource> {
-        private final AsyncCallableWithConnection callable;
-        AsyncCallableWithConnectionCallback(final AsyncCallableWithConnection callable) {
-            this.callable = callable;
-        }
-        @Override
-        public void onResult(@Nullable final AsyncConnectionSource source, @Nullable final Throwable t) {
-            if (t != null) {
-                callable.call(null, t);
-            } else {
-                withAsyncConnectionSourceCallableConnection(assertNotNull(source), callable);
-            }
-        }
-    }
-
-    private static class AsyncCallableWithSourceCallback implements SingleResultCallback<AsyncConnectionSource> {
-        private final AsyncCallableWithSource callable;
-        AsyncCallableWithSourceCallback(final AsyncCallableWithSource callable) {
-            this.callable = callable;
-        }
-        @Override
-        public void onResult(@Nullable final AsyncConnectionSource source, @Nullable final Throwable t) {
-            if (t != null) {
-                callable.call(null, t);
-            } else {
-                withAsyncConnectionSource(assertNotNull(source), callable);
-            }
-        }
-    }
-
-    private static void withAsyncConnectionSourceCallableConnection(final AsyncConnectionSource source,
-                                                                    final AsyncCallableWithConnection callable) {
-        source.getConnection((connection, t) -> {
-            source.release();
-            if (t != null) {
-                callable.call(null, t);
-            } else {
-                callable.call(connection, null);
-            }
-        });
-    }
-
-    private static void withAsyncConnectionSource(final AsyncConnectionSource source, final AsyncCallableWithSource callable) {
-        callable.call(source, null);
-    }
-
-    private static void withAsyncConnectionSource(final AsyncConnectionSource source, final AsyncCallableWithConnectionAndSource callable) {
-        source.getConnection((result, t) -> callable.call(source, result, t));
-    }
-
-    private static class AsyncCallableWithConnectionAndSourceCallback implements SingleResultCallback<AsyncConnectionSource> {
-        private final AsyncCallableWithConnectionAndSource callable;
-
-        AsyncCallableWithConnectionAndSourceCallback(final AsyncCallableWithConnectionAndSource callable) {
-            this.callable = callable;
-        }
-
-        @Override
-        public void onResult(@Nullable final AsyncConnectionSource source, @Nullable final Throwable t) {
-            if (t != null) {
-                callable.call(null, null, t);
-            } else {
-                withAsyncConnectionSource(assertNotNull(source), callable);
-            }
-        }
-    }
-
-    private OperationHelper() {
-    }
 
     /**
      * This internal exception is used to
      * <ul>
-     *     <li>on one hand allow propagating exceptions from {@link #withSuppliedResource(Supplier, boolean, Function)} /
-     *     {@link #withAsyncSuppliedResource(AsyncCallbackSupplier, boolean, SingleResultCallback, AsyncCallbackFunction)} and similar
-     *     methods so that they can be properly retried, which is useful, e.g.,
+     *     <li>on one hand allow propagating exceptions from {@link SyncOperationHelper#withSuppliedResource(Supplier, boolean, Function)} /
+     *     {@link AsyncOperationHelper#withAsyncSuppliedResource(AsyncCallbackSupplier, boolean, SingleResultCallback, AsyncCallbackFunction)}
+     *     and similar methods so that they can be properly retried, which is useful, e.g.,
      *     for {@link com.mongodb.MongoConnectionPoolClearedException};</li>
      *     <li>on the other hand to prevent them from propagation once the retry decision is made.</li>
      * </ul>
      *
-     * @see #withSuppliedResource(Supplier, boolean, Function)
-     * @see #withAsyncSuppliedResource(AsyncCallbackSupplier, boolean, SingleResultCallback, AsyncCallbackFunction)
+     * @see SyncOperationHelper#withSuppliedResource(Supplier, boolean, Function)
+     * @see AsyncOperationHelper#withAsyncSuppliedResource(AsyncCallbackSupplier, boolean, SingleResultCallback, AsyncCallbackFunction)
      */
-    static final class ResourceSupplierInternalException extends RuntimeException {
+    public static final class ResourceSupplierInternalException extends RuntimeException {
         private static final long serialVersionUID = 0;
 
-        private ResourceSupplierInternalException(final Throwable cause) {
+        ResourceSupplierInternalException(final Throwable cause) {
             super(assertNotNull(cause));
         }
 
@@ -521,5 +245,8 @@ final class OperationHelper {
         public Throwable getCause() {
             return assertNotNull(super.getCause());
         }
+    }
+
+    private OperationHelper() {
     }
 }
