+++ driver-core/src/test/unit/com/mongodb/internal/operation/CommandOperationHelperSpecification.groovy
@@ -17,43 +17,15 @@
 package com.mongodb.internal.operation
 
 import com.mongodb.MongoCommandException
-import com.mongodb.MongoWriteConcernException
-import com.mongodb.ReadConcern
-import com.mongodb.ReadPreference
 import com.mongodb.ServerAddress
-import com.mongodb.connection.ConnectionDescription
-import com.mongodb.connection.ServerDescription
-import com.mongodb.connection.ServerType
-import com.mongodb.internal.async.SingleResultCallback
-import com.mongodb.internal.binding.AsyncConnectionSource
-import com.mongodb.internal.binding.AsyncReadBinding
-import com.mongodb.internal.binding.AsyncWriteBinding
-import com.mongodb.internal.binding.ConnectionSource
-import com.mongodb.internal.binding.ReadBinding
-import com.mongodb.internal.binding.WriteBinding
-import com.mongodb.internal.connection.AsyncConnection
-import com.mongodb.internal.connection.Connection
-import com.mongodb.internal.session.SessionContext
-import com.mongodb.internal.validator.NoOpFieldNameValidator
 import org.bson.BsonBoolean
 import org.bson.BsonDocument
 import org.bson.BsonInt32
-import org.bson.BsonNull
 import org.bson.BsonString
-import org.bson.codecs.BsonDocumentCodec
-import org.bson.codecs.Decoder
 import spock.lang.Specification
 
-import static com.mongodb.ReadPreference.primary
-import static com.mongodb.internal.operation.CommandOperationHelper.executeCommand
-import static com.mongodb.internal.operation.CommandOperationHelper.executeCommandAsync
-import static com.mongodb.internal.operation.CommandOperationHelper.executeRetryableRead
-import static com.mongodb.internal.operation.CommandOperationHelper.executeRetryableReadAsync
-import static com.mongodb.internal.operation.CommandOperationHelper.executeRetryableWrite
-import static com.mongodb.internal.operation.CommandOperationHelper.executeRetryableWriteAsync
 import static com.mongodb.internal.operation.CommandOperationHelper.isNamespaceError
 import static com.mongodb.internal.operation.CommandOperationHelper.rethrowIfNotNamespaceError
-import static com.mongodb.internal.operation.OperationUnitSpecification.getMaxWireVersionForServerVersion
 
 class CommandOperationHelperSpecification extends Specification {
 
@@ -116,218 +88,6 @@ class CommandOperationHelperSpecification extends Specification {
         rethrowIfNotNamespaceError(new MongoCommandException(new BsonDocument('ok', BsonBoolean.FALSE)
                                                                      .append('code', new BsonInt32(26)),
                                                              new ServerAddress()), 'some value') == 'some value'
-    }
-
-    def 'should set read preference to primary to false when using WriteBinding'() {
-        given:
-        def dbName = 'db'
-        def command = new BsonDocument()
-        def decoder = Stub(Decoder)
-        def connection = Mock(Connection)
-        def function = Stub(CommandOperationHelper.CommandWriteTransformer)
-        def connectionSource = Stub(ConnectionSource) {
-            getServerApi() >> null
-            getConnection() >> connection
-        }
-        def writeBinding = Stub(WriteBinding) {
-            getServerApi() >> null
-            getWriteConnectionSource() >> connectionSource
-        }
-        def connectionDescription = Stub(ConnectionDescription)
-
-        when:
-        executeCommand(writeBinding, dbName, command, decoder, function)
-
-        then:
-        _ * connection.getDescription() >> connectionDescription
-        1 * connection.command(dbName, command, _, primary(), decoder, writeBinding) >> new BsonDocument()
-        1 * connection.release()
-    }
-
-    def 'should retry with retryable exception'() {
-        given:
-        def dbName = 'db'
-        def command = BsonDocument.parse('''{findAndModify: "coll", query: {a: 1}, new: false, update: {$inc: {a :1}}, txnNumber: 1}''')
-        def commandCreator = { serverDescription, connectionDescription -> command }
-        def decoder = new BsonDocumentCodec()
-        def results = [
-            BsonDocument.parse('{ok: 1.0, writeConcernError: {code: 91, errmsg: "Replication is being shut down"}}'),
-            BsonDocument.parse('{ok: 1.0, writeConcernError: {code: -1, errmsg: "UnknownError"}}')] as Queue
-        def connection = Mock(Connection) {
-            _ * release()
-            _ * getDescription() >> Stub(ConnectionDescription) {
-                getMaxWireVersion() >> getMaxWireVersionForServerVersion([4, 0, 0])
-                getServerType() >> ServerType.REPLICA_SET_PRIMARY
-            }
-        }
-        def connectionSource = Stub(ConnectionSource) {
-            _ * getConnection() >> connection
-            _ * getServerDescription() >> Stub(ServerDescription) {
-                getLogicalSessionTimeoutMinutes() >> 1
-            }
-        }
-        def writeBinding = Stub(WriteBinding) {
-            getWriteConnectionSource() >> connectionSource
-            getServerApi() >> null
-            getSessionContext() >> Stub(SessionContext) {
-                hasSession() >> true
-                hasActiveTransaction() >> false
-                getReadConcern() >> ReadConcern.DEFAULT
-            }
-        }
-
-        when:
-        executeRetryableWrite(writeBinding, dbName, primary(), new NoOpFieldNameValidator(), decoder, commandCreator,
-                FindAndModifyHelper.transformer()) { cmd -> cmd }
-
-        then:
-        2 * connection.command(dbName, command, _, primary(), decoder, writeBinding) >> { results.poll() }
-
-        then:
-        def ex = thrown(MongoWriteConcernException)
-        ex.writeConcernError.code == -1
-    }
-
-    def 'should retry with retryable exception async'() {
-        given:
-        def dbName = 'db'
-        def command = BsonDocument.parse('''{findAndModify: "coll", query: {a: 1}, new: false, update: {$inc: {a :1}}, txnNumber: 1}''')
-        def serverDescription = Stub(ServerDescription)
-        def connectionDescription = Stub(ConnectionDescription) {
-            getMaxWireVersion() >> getMaxWireVersionForServerVersion([4, 0, 0])
-            getServerType() >> ServerType.REPLICA_SET_PRIMARY
-        }
-        def commandCreator = { serverDesc, connectionDesc -> command }
-        def callback = new SingleResultCallback() {
-            def result
-            def throwable
-            @Override
-            void onResult(final Object result, final Throwable t) {
-                this.result = result
-                this.throwable = t
-            }
-        }
-        def decoder = new BsonDocumentCodec()
-        def results = [
-                BsonDocument.parse('{ok: 1.0, writeConcernError: {code: 91, errmsg: "Replication is being shut down"}}'),
-                BsonDocument.parse('{ok: 1.0, writeConcernError: {code: -1, errmsg: "UnknownError"}}')] as Queue
-
-        def connection = Mock(AsyncConnection) {
-            _ * getDescription() >> connectionDescription
-        }
-
-        def connectionSource = Stub(AsyncConnectionSource) {
-            getServerApi() >> null
-            getConnection(_) >> { it[0].onResult(connection, null) }
-            _ * getServerDescription() >> serverDescription
-        }
-        def asyncWriteBinding = Stub(AsyncWriteBinding) {
-            getServerApi() >> null
-            getWriteConnectionSource(_) >> { it[0].onResult(connectionSource, null) }
-            getSessionContext() >> Stub(SessionContext) {
-                hasSession() >> true
-                hasActiveTransaction() >> false
-                getReadConcern() >> ReadConcern.DEFAULT
-            }
-        }
-
-        when:
-        executeRetryableWriteAsync(asyncWriteBinding, dbName, primary(), new NoOpFieldNameValidator(), decoder,
-                commandCreator, FindAndModifyHelper.asyncTransformer(), { cmd -> cmd }, callback)
-
-        then:
-        2 * connection.commandAsync(dbName, command, _, primary(), decoder, *_) >> { it.last().onResult(results.poll(), null) }
-
-        then:
-        callback.throwable instanceof MongoWriteConcernException
-        callback.throwable.writeConcernError.code == -1
-    }
-
-    def 'should use the ConnectionSource readPreference'() {
-        given:
-        def dbName = 'db'
-        def command = new BsonDocument('fakeCommandName', BsonNull.VALUE)
-        def commandCreator = { serverDescription, connectionDescription -> command }
-        def decoder = Stub(Decoder)
-        def function = Stub(CommandOperationHelper.CommandReadTransformer)
-        def connection = Mock(Connection)
-        def connectionSource = Stub(ConnectionSource) {
-            getConnection() >> connection
-            getReadPreference() >> readPreference
-        }
-        def readBinding = Stub(ReadBinding) {
-            getReadConnectionSource() >> connectionSource
-            getServerApi() >> null
-        }
-        def connectionDescription = Stub(ConnectionDescription)
-
-        when:
-        executeRetryableRead(readBinding, dbName, commandCreator, decoder, function, false)
-
-        then:
-        _ * connection.getDescription() >> connectionDescription
-        1 * connection.command(dbName, command, _, readPreference, decoder, readBinding) >> new BsonDocument()
-        1 * connection.release()
-
-        where:
-        readPreference << [primary(), ReadPreference.secondary()]
-    }
-
-    def 'should set read preference to primary when using AsyncWriteBinding'() {
-        given:
-        def dbName = 'db'
-        def command = new BsonDocument()
-        def callback = Stub(SingleResultCallback)
-        def connection = Mock(AsyncConnection)
-        def connectionSource = Stub(AsyncConnectionSource) {
-            getServerApi() >> null
-            getConnection(_) >> { it[0].onResult(connection, null) }
-        }
-        def asyncWriteBinding = Stub(AsyncWriteBinding) {
-            getServerApi() >> null
-            getWriteConnectionSource(_) >> { it[0].onResult(connectionSource, null) }
-        }
-        def connectionDescription = Stub(ConnectionDescription)
-
-        when:
-        executeCommandAsync(asyncWriteBinding, dbName, command, connection, { t, conn -> t }, callback)
-
-        then:
-        _ * connection.getDescription() >> connectionDescription
-        1 * connection.commandAsync(dbName, command, _, primary(), *_) >> { it.last().onResult(1, null) }
-//        1 * connection.release()
-    }
-
-    def 'should use the AsyncConnectionSource readPreference'() {
-        given:
-        def dbName = 'db'
-        def command = new BsonDocument('fakeCommandName', BsonNull.VALUE)
-        def commandCreator = { serverDescription, connectionDescription -> command }
-        def decoder = Stub(Decoder)
-        def callback = Stub(SingleResultCallback)
-        def function = Stub(CommandOperationHelper.CommandReadTransformerAsync)
-        def connection = Mock(AsyncConnection)
-        def connectionSource = Stub(AsyncConnectionSource) {
-            getServerApi() >> null
-            getConnection(_) >> { it[0].onResult(connection, null) }
-            getReadPreference() >> readPreference
-        }
-        def asyncReadBinding = Stub(AsyncReadBinding) {
-            getServerApi() >> null
-            getReadConnectionSource(_)  >> { it[0].onResult(connectionSource, null) }
-        }
-        def connectionDescription = Stub(ConnectionDescription)
-
-        when:
-        executeRetryableReadAsync(asyncReadBinding, dbName, commandCreator, decoder, function, false, callback)
-
-        then:
-        _ * connection.getDescription() >> connectionDescription
-        1 * connection.commandAsync(dbName, command, _, readPreference, decoder, *_) >> { it.last().onResult(1, null) }
-        1 * connection.release()
-
-        where:
-        readPreference << [primary(), ReadPreference.secondary()]
     }
 
 }
